+++ libcloud/test/compute/test_cloudstack.py
@@ -150,6 +150,31 @@ class CloudStackCommonTestCase(TestCaseMixin):
         self.assertEqual(node.extra['image_id'], image.id)
         self.assertEqual(node.private_ips[0], ipaddress)
 
+    def test_create_node_ex_rootdisksize(self):
+        CloudStackMockHttp.fixture_tag = 'rootdisksize'
+        size = self.driver.list_sizes()[0]
+        image = self.driver.list_images()[0]
+        location = self.driver.list_locations()[0]
+        volumes = self.driver.list_volumes()
+        rootdisksize = '50'
+
+        networks = [nw for nw in self.driver.ex_list_networks()
+                    if str(nw.zoneid) == str(location.id)]
+
+        node = self.driver.create_node(name='rootdisksize',
+                                       location=location,
+                                       image=image,
+                                       size=size,
+                                       networks=networks,
+                                       ex_rootdisksize=rootdisksize)
+        self.assertEqual(node.name, 'rootdisksize')
+        self.assertEqual(node.extra['size_id'], size.id)
+        self.assertEqual(node.extra['zone_id'], location.id)
+        self.assertEqual(node.extra['image_id'], image.id)
+        self.assertEqual(1, len(volumes))
+        self.assertEqual('ROOT-69941', volumes[0].name)
+        self.assertEqual(53687091200, volumes[0].size)
+
     def test_create_node_ex_start_vm_false(self):
         CloudStackMockHttp.fixture_tag = 'stoppedvm'
         size = self.driver.list_sizes()[0]